Apply localstorage for the editor.
The vue instance now fetches the local storage object on load
The save button in the html is a temporary method for saving
content.

Eventually, saving should be automatic, and occur on close.

Also, I set up constants to avoid spelling / syntax / because I needed
to be superfluous about something.

listening to: Memoryhouse - Bonfire